fix: fixed inline css & function
