+++ apps/web/components/Feature/Settings/Export/index.tsx
@@ -3,21 +3,28 @@
 import React, { useState } from "react";
 import { MonoText } from "@/components/UI/Monotext";
 import { useTranslation } from "react-i18next";
-import { Card, CardTop, TextBlock, Section, FieldBox, Row } from "./styles";
+import { Card, CardTop, TextBlock, Section, FieldBox } from "./styles";
 import { SETTINGS } from "@/utils/translationKeys";
 import DropdownField from "@/components/UI/InputFields/DropdownField";
 import DateRangeField from "@/components/UI/InputFields/DateRangeField";
 import GenericButton from "@/components/UI/GenericButton";
 import { VARIANT } from "@/utils/Constants";
 import COLORS from "@repo/ui/colors";
+import { EXPORT_TYPE_OPTIONS } from "@/utils/exportOptions";
 
 export default function ExportContent() {
   const { t } = useTranslation();
-  const [includeMedia, setIncludeMedia] = useState(true);
+
   const [exportType, setExportType] = useState<string>("users-email-signups");
+  const [startDate, setStartDate] = useState<string>("");
+  const [endDate, setEndDate] = useState<string>("");
 
   const handleRequest = () => {
-    console.log("request export", { includeMedia });
+    console.log("request export", {
+      exportType,
+      startDate,
+      endDate,
+    });
   };
 
   return (
@@ -40,14 +47,7 @@ export default function ExportContent() {
         <FieldBox>
           <DropdownField
             label={t(SETTINGS.export.typeLabel)}
-            options={[
-              {
-                value: "users-email-signups",
-                label: t(SETTINGS.export.exportTypeUsersEmailSignups),
-              },
-              { value: "sales", label: t(SETTINGS.export.exportTypeSales) },
-              { value: "views", label: t(SETTINGS.export.exportTypeViews) },
-            ]}
+            options={EXPORT_TYPE_OPTIONS(t)}
             value={exportType}
             onChange={(v) => setExportType(v)}
           />
@@ -56,10 +56,10 @@ export default function ExportContent() {
         <FieldBox>
           <DateRangeField
             label={t(SETTINGS.export.dateLabel)}
-            start={""}
-            end={""}
-            onChangeStart={() => {}}
-            onChangeEnd={() => {}}
+            start={startDate}
+            end={endDate}
+            onChangeStart={setStartDate}
+            onChangeEnd={setEndDate}
           />
         </FieldBox>
       </Section>
