+++ apps/web/components/UI/InputFields/styles.ts
@@ -296,96 +296,6 @@ export const DatePopupActions = styled.div`
   gap: 12px;
 `;
 
-export const CalendarWrapper = styled.div`
-  display: flex;
-  gap: 24px;
-  align-items: flex-start;
-  position: relative;
-
-  &::before {
-    content: "";
-    position: absolute;
-    left: 50%;
-    top: 0;
-    bottom: 0;
-    width: 1px;
-    background: ${({ theme }) => theme.colors.neutral.GRAY_200};
-    transform: translateX(-50%);
-  }
-`;
-
-export const CalendarMonth = styled.div`
-  min-width: 300px;
-  flex: 1;
-`;
-
-export const CalendarHeader = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 12px;
-`;
-
-export const MonthTitle = styled.div`
-  ${({ theme }) => theme.typography.Body_SemiBold};
-  color: ${({ theme }) => theme.colors.primary.BLACK};
-`;
-
-export const WeekDays = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  gap: 6px;
-  margin-bottom: 8px;
-  color: ${({ theme }) => theme.colors.neutral.GRAY};
-  ${({ theme }) => theme.typography.Body_Regular};
-`;
-
-export const DaysGrid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-  gap: 6px;
-`;
-
-export const DayButton = styled.button<{ $isOutside?: boolean }>`
-  width: 100%;
-  height: 40px;
-  width: 40px;
-  border-radius: 50%;
-  border: none;
-  background: ${({ $isOutside, theme }) =>
-    $isOutside ? theme.colors.neutral.GRAY_100 : "transparent"};
-  cursor: pointer;
-  ${({ theme }) => theme.typography.Body_Regular};
-  color: ${({ theme }) => theme.colors.primary.BLACK};
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-export const DaySelected = styled.div`
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background: ${({ theme }) => theme.colors.primary.BLACK};
-  color: ${({ theme }) => theme.colors.primary.WHITE};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  ${({ theme }) => theme.typography.Body_Regular};
-`;
-
-export const DayInRange = styled.div`
-  min-width: 40px;
-  height: 40px;
-  border-radius: 999px;
-  background: ${({ theme }) => theme.colors.neutral.GRAY_200};
-  color: ${({ theme }) => theme.colors.primary.BLACK};
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  ${({ theme }) => theme.typography.Body_Regular};
-`;
-
 export const CalendarNavButton = styled.button`
   width: 36px;
   height: 36px;
