+++ apps/web/utils/formatDate.ts
@@ -1,0 +1,16 @@
+export function formatDate(iso?: string) {
+  if (!iso) return "";
+
+  try {
+    const d = new Date(iso);
+    if (isNaN(d.getTime())) return "";
+
+    const dd = String(d.getDate()).padStart(2, "0");
+    const mm = String(d.getMonth() + 1).padStart(2, "0");
+    const yyyy = d.getFullYear();
+
+    return `${dd}.${mm}.${yyyy}`;
+  } catch {
+    return "";
+  }
+}
